requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
get==2019.4.13
idna==2.8
Markdown==3.1.1
post==2019.4.13
psycopg2==2.8.4
public==2019.4.13
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
